+++ src/yafowil/widget/dynatree/__init__.py
@@ -1,4 +1,5 @@
 from yafowil.base import factory
+from yafowil.utils import entry_point
 import os
 
 
@@ -32,6 +33,7 @@ bootstrap_css = [{
 }]
 
 
+@entry_point(order=10)
 def register():
     import widget
     factory.register_theme('default', 'yafowil.widget.dynatree',