+++ HISTORY.rst
@@ -5,14 +5,16 @@ History
 1.4.2dev
 --------
 
-- No notable changes yet
-  [rnix, 2014-04-30]
+- Bootstrap 3 compatibility.
+  [rnix, 2014-08-06]
+
 
 1.4.1
 -----
 
 - use ``yafowil.utils.attr_value`` wherever possible.
   [rnix, 2012-10-25]
+
 
 1.4
 ---
@@ -23,11 +25,13 @@ History
 - Remove example app
   [rnix, 2012-06-12]
 
+
 1.3
 ---
 
 - make resources available the new way and depend on YAFOWIL 1.3
   [jensens, 2012-02-19]
+
 
 1.2
 ---
@@ -38,11 +42,13 @@ History
 - make ready for z2c.autoinclude+Plone (only if available).
   [jensens, 2011-09-20]
 
+
 1.1
 ---
 
 - adopt to yafowil 1.1
   [rnix, 2011-07-08]
+
 
 1.0.2
 -----
@@ -50,11 +56,13 @@ History
 - fixed self closing divs, which are not allowed
   [agitator, 2001-06-03]
 
+
 1.0.1
 -----
 
 - test coverage
   [rnix, 2011-05-07]
+
 
 1.0
 ---
@@ -64,6 +72,7 @@ History
 
 - js event trigger onSelect
   [jensens, 2011-01-18]
+
 
 1.0-beta
 --------